backend/requirements.txt
fastapi==0.109.0
uvicorn==0.23.2
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
httpx==0.26.0
groq>=0.4.0
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.18.0
huggingface-hub>=0.20.0
pydantic-settings==2.1.0
bcrypt==3.2.0
cryptography==42.0.1
email-validator
razorpay==1.4.1
pypdf
openpyxl
Pillow
pandas
fpdf
qrcode
requests